requirements.txt
pandas
streamlit
google-generativeai
pandas
python-dotenv
langchain
PyPDF2
chromadb
langchain_google_genai
faiss-cpu
langchain-community